requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
sendgrid>=6.11.0
lxml>=5.1.0

# Optional: Keepa API for reliable price data (upgrades from CCC scraper)
# Sign up at https://keepa.com/#!api (~$20/month)
# Install with: pip install keepa
keepa>=1.3.0
numpy>=1.24.0  # Required by keepa
